I have updated the urls and the working of the topic-related activities. Now they all take place in a shared controller for all 3 roles.
I have also added the working of the organization member role. It does exactly the same things as parent or student.

+++ src/main/java/com/adtimokhin/controllers/main/TopicController.java
@@ -1,5 +1,6 @@
 package com.adtimokhin.controllers.main;
 
+import com.adtimokhin.enums.Role;
 import com.adtimokhin.models.comment.Comment;
 import com.adtimokhin.models.topic.Topic;
 import com.adtimokhin.models.user.User;
@@ -20,12 +21,15 @@ import java.util.List;
 
 /**
  * @author adtimokhin
- * 22.04.2021
+ * 09.06.2021
  **/
 
 @Controller
-@RequestMapping("student")
-public class StudentController {
+public class TopicController {
+
+
+    @Autowired
+    private ContextProvider contextProvider;
 
     @Autowired
     private TopicService topicService;
@@ -43,35 +47,36 @@ public class StudentController {
     private ReportService reportService;
 
     @Autowired
-    private ContextProvider contextProvider;
-
-    @Autowired
     private CommentTagsServiceImpl tagsService;
-
 
     @GetMapping("/topics")
     public String getAllTopics(Model model) {
-        model.addAttribute("topics", topicService.getAllTopicsForStudents());
-        return "/student/studentForumPage";
+        if (isStudent()){
+            model.addAttribute("topics", topicService.getAllTopicsForStudents());
+        }else {
+            model.addAttribute("topics", topicService.getAllTopics());
+        }
+        return "main/pages/forumPage";
     }
+
 
     @GetMapping("/topic/{id}")
     public String getTopic(@PathVariable(name = "id") long id, Model model) {
         Topic topic = topicService.getTopic(id);
         if (topic == null) {
             // TODO: throw a 404 error
-            return "redirect:/student/topics";
+            return "redirect:/topics";
         }
-        if (!topicService.isUserAllowedOntoTopic(topic)) {
-            return "redirect:/student/topics";//Todo:redirect to some error page (access denied page, e.g.)
+        if (!topicService.isUserAllowedOntoTopic(topic, contextProvider.getUser())) {
+            return "redirect:/topics";//Todo:redirect to some error page (access denied page, e.g.)
         }
         List<Comment> comments = commentService.getAllCommentsByTopic(id);
         List<Comment> flaggedComments = commentService.getFlaggedComment(comments);
         if (flaggedComments != null){
             model.addAttribute("flaggedComments" , flaggedComments);
         }
-        User u = contextProvider.getUser();
-        List<Long> likedCommentIds = likeService.getAllLikedCommentIdsByUser(u, comments);
+        User user = contextProvider.getUser();
+        List<Long> likedCommentIds = likeService.getAllLikedCommentIdsByUser(user, comments);
         model.addAttribute("likedComments", likedCommentIds);
         model.addAttribute("topic", topic);
         model.addAttribute("comments", comments);
@@ -79,7 +84,7 @@ public class StudentController {
 
         //some special functionality is only available to a user that have initiated the topic.
         // We need to check if a user that gets the view is the same user that have created the topic
-        if(topicService.isUserCreatedTopic(topic, u)){
+        if(topicService.isUserCreatedTopic(topic, user)){
             model.addAttribute("theCreator" , true);
         }else {
             model.addAttribute("theCreator", false);
@@ -90,7 +95,7 @@ public class StudentController {
         }else {
             model.addAttribute("closed" , false);
         }
-        return "/student/studentTopicPage";
+        return "/main/pages/topicPage";
     }
 
 
@@ -100,17 +105,17 @@ public class StudentController {
     @GetMapping("/add/topic")
     public String studentAddTopic(Model model) {
         model.addAttribute("topic", new Topic());
-        return "/student/studentNewTopicPage";
+        return "/main/pages/newTopicPage";
     }
 
     @PostMapping("/add/topic")
     public String studentAddTopic(@ModelAttribute Topic topic, BindingResult result) {
         topicValidator.validate(topic, result);
         if (result.hasErrors()) {
-            return "/student/studentNewTopicPage";
+            return "/main/pages/newTopicPage"; // Todo: have a look at how errors are displayed.
         }
         topicService.addTopic(topic);
-        return "redirect:/student/topics";
+        return "redirect:/topics";
 
     }
 
@@ -118,7 +123,7 @@ public class StudentController {
     public String closeTopic(@RequestParam(name = "topicId") long topicId){
         User user = contextProvider.getUser();
         topicService.closeTopic(topicId, user);
-        return "redirect:/student/topics";
+        return "redirect:/topics";
     }
 
 
@@ -126,7 +131,7 @@ public class StudentController {
     public String openTopic(@RequestParam(name = "topicId") long topicId){
         User user = contextProvider.getUser();
         topicService.openTopic(topicId, user);
-        return "redirect:/student/topics";
+        return "redirect:/topics";
     }
 
     //working with comments
@@ -137,33 +142,45 @@ public class StudentController {
                                     @RequestParam(name = "topicId") long topicId,
                                     @RequestParam(name = "tags", required = false) List<Long> tags) {
 
-        if (!topicService.isUserAllowedOntoTopic(topicService.getTopic(topicId))) {
-            return "redirect:/student/topics"; //Todo:redirect to some error page (access denied page, e.g.)
+        if (!topicService.isUserAllowedOntoTopic(topicService.getTopic(topicId) , contextProvider.getUser())) {
+            return "redirect:/topics"; //Todo:redirect to some error page (access denied page, e.g.)
         }
 
         commentService.addComment(msg, topicId, tags);
-        return "redirect:/student/topic/" + topicId;
+        return "redirect:/topic/" + topicId;
     }
 
     @PostMapping("/update/comment/flag")
     public String flagComment(@RequestParam(name = "commentId") long commentId){
         commentService.flagComment(commentId, contextProvider.getUser());
-        return "redirect:/student/topics";
+        return "redirect:/topics";
     }
     // add a like to a comment
     @PostMapping("/add/like")
     public String studentAddLike(@RequestParam(name = "comment") long commentId) {
-        Comment comment = commentService.getCommentById(commentId); // Todo: это обращение к бд - ненужное и зря нагружает систему. Нужно это как-то пофиксить
+        Comment comment = commentService.getCommentById(commentId);
+        if (comment == null){
+            return "redirect:/topics"; //Todo:redirect to some error page (access denied page, e.g.)
+        }
+        if (!topicService.isUserAllowedOntoTopic(comment.getTopic() , contextProvider.getUser())) {
+            return "redirect:/topics"; //Todo:redirect to some error page (access denied page, e.g.)
+        }
         likeService.addLike(contextProvider.getUser(), comment);
-        return "redirect:/student/topics";
+        return "redirect:/topics";
     }
 
     // removing a like from a comment
     @PostMapping("/remove/like")
     public String studentRemoveLike(@RequestParam(name = "comment") long commentId) {
-        Comment comment = commentService.getCommentById(commentId); // Todo: это обращение к бд - ненужное и зря нагружает систему. Нужно это как-то пофиксить
+        Comment comment = commentService.getCommentById(commentId);
+        if (comment == null){
+            return "redirect:/topics"; //Todo:redirect to some error page (access denied page, e.g.)
+        }
+        if (!topicService.isUserAllowedOntoTopic(comment.getTopic() , contextProvider.getUser())) {
+            return "redirect:/topics"; //Todo:redirect to some error page (access denied page, e.g.)
+        }
         likeService.deleteLike(contextProvider.getUser(), comment);
-        return "redirect:/student/topics";
+        return "redirect:/topics";
     }
 
     // working with reports and bans
@@ -174,8 +191,28 @@ public class StudentController {
             @RequestParam(name = "reportedUserId") long reportedUserId,
             @RequestParam(name = "causeId") long causeId
     ){
-
+        // only is user is allowed on the topic should he/she create a report.
+        if (isComment){
+            Comment comment = commentService.getCommentById(commentOrTopicId);
+            if (!topicService.isUserAllowedOntoTopic(comment.getTopic() , contextProvider.getUser())){
+                return "redirect:/topics"; //Todo:redirect to some error page (access denied page, e.g.)
+            }
+        }else {
+            Topic topic = topicService.getTopic(commentOrTopicId);
+            if (!topicService.isUserAllowedOntoTopic(topic, contextProvider.getUser())){
+                return "redirect:/topics"; //Todo:redirect to some error page (access denied page, e.g.)
+            }
+        }
         reportService.addReport(commentOrTopicId , isComment, reportedUserId, contextProvider.getUser().getId(), causeId);
-        return "redirect:/student/topics";
+        return "redirect:/topics";
+    }
+    private Role getRole() {
+        return (Role) contextProvider.getUser().getRoles().toArray()[0];
+    }
+
+    //TODO: add to config files that now all urls associated with topics can be entered either by STUDENT PARENT and MEMBER.
+
+    private boolean isStudent() {
+        return getRole().equals(Role.ROLE_STUDENT);
     }
 }
